Move COM stuff to separate module
